+++ tests/scatter_speed.py
@@ -24,8 +24,7 @@ ptcl_spacing = 1.  # Lagrangian space Cartesian particle grid spacing, in Mpc/h 
 ptcl_grid_shape = (ngrid,) * 3
 conf = Configuration(ptcl_spacing, ptcl_grid_shape, mesh_shape=1.,float_dtype=jnp.float32)  # 1x mesh shape
 ptcl = Particles.gen_grid(conf)
-pmid = ptcl.pmid.astype(jnp.uint32)
-#pmid = ptcl.pmid
+pmid = ptcl.pmid
 mesh = random.uniform(key, shape=conf.mesh_shape,dtype=jnp.float32,minval=0.0,maxval=2.0)
 val = random.uniform(key, shape=(ngrid*ngrid*ngrid,),dtype=jnp.float32,minval=0.0,maxval=2.0)
 disp = random.uniform(key, shape=(ngrid*ngrid*ngrid,3),dtype=jnp.float32,minval=0.0,maxval=2.0)
